fix(Makefile): does not create package scripts
Description: The command `make create-package` does not show the
script folder name when asking if it needs to create it. Also does not
create it. If the user selects to create the script folder, the Makefile
creates a '<script folder name>/scripts.stow/' directory (create a new
package with the name of the script).

Cause: the `create-package` target in the Makefile calls the
`create_script_folder` recipe. This recipe requires two arguments:

* $(1): The package name (as string)
* $(2): The script type (as string)

But the call to this recipe in the `create-package` target only provides
the $(2) parameter, that will be interpreted as the package name by the
recipe

Solution: add '${PACKAGE}' as $(1), and the script folder name as $(2)