feat(tmux): set options required to use Neovim
Set 'escape-time' and 'focus-events' options. The first defines the
delay that differentiates between a ESC key or a key with special
meaning (alt, function key, etc.). Whiteout this, there will be a long
delay before exiting insert mode. The second is necessary for Neovim to
be able to raise the 'FocusGained' and 'FocusLost' events correctly