fix(bin): `chdesk` create simlink into Desktop
Bug: if the user does not initialize the `chdesk` utility properly, it
creates a symbolic link of the selected desktop (with the `chdesk change`
command) inside the '~/Desktop' folder. The desired behavior is to throw
an error

Cause: the `chdesk` utility does not checks if the '~/Desktop' folder is
a symbolic link or a normal folder

Solution: add the `-T` option to the `ln` command inside the `chdesk
change` command. This makes the `ln` command treat the '~/Desktop'
folder as an file and raise an error if it is not a symbolic link