fix(nvim): error when using S-TAB in insert mode
Bug: press S-TAB in insert mode causes an error with this message:
"attempt to call global 'fallback' (a nil value)"

Cause: the `nvim-cmp` map at line 239 maps the S-TAB to a function
that must receive a `fallback` argument. But the previous implementation
of this function does not has this argument. Every time the user press
the S-TAB in insert mode, this mapping will attempt to call a fallback
function that does not exist

Solution: add the `fallback` argument to the mapping function