fix(nvim): CTRL + BACKSPACE key mapping
Bug: CTRL + BACKSPACE does not work if Neovim is opened in Kitty. CTRL +
BACKSPACE moves the cursor right

Cause: Kitty uses <C-BS> to emulate CTRL + BACKSPACE. The old
implementation considers that the terminal emulator sends a <C-h>
instead <C-BS>. This behavior depends on the terminal emulator. Some use
<C-h> or <C-w> instead of <C-BS>. The last implementation of <C-BS>
moves the cursor right because the delete operation is performed with a
normal command. If the cursor is in the last column, the delete
operation places the cursor one column before the last (moves one column
to the left)

Solution: map the <C-BS> to execute the delete operation. Map the <C-h>
and <C-w> to be equivalent to <C-BS>, so other terminal emulators will
also execute the delete operation if the user pressed CTRL + BACKSPACE.
To fix the cursor movement error, we need to move the cursor to
right only if it is in the last column