fix(nvim): can not highlight white space error
Bug: the white space highlight utility does not works if the user open
the buffer in a window whose last buffer should not highlight them. E.g.
You open a help menu (should not highlight white space errors) and,
after this, open a file in the same window. The desired behavior is t
enable the white space error highlight. But the old configuration does
not enables it

Cause: set the `__whitespace_error_loaded` flag before checking if
should load the white space errors highlights. This enables the flag even
if the highlights are not applied

Solution: set the `__whitespace_error_loaded` flag after checking if
should load the highlights in the current window