feat(nvim): add `noice.nvim` plugin
Remove the 'ray-x/lsp_signature.nvim' plugin. `noice` can replace the
features of it that I use

`noice` redirects all the printed text with the `print()` function to
the notifications. The notifications at the top may hide the start of
the file. To avoid this situation, show the notification from bottom to
top

Add Tree-sitter 'regex' parser. Required by `noice.nvim`

Show the register used to save the current typing macro in the status
line because `noice.nvim` removes the default line used to show it