feat(bin): synced version of some scripts
Add a synced version of these scripts:

* default_file_manager

* default_term

* nvim_new_win

The synced version have the same name of the original one, but with the
'_synced' suffix. These new scripts run the same program as its
counterpart, but in the current terminal instance. If you close the
terminal instance that started the script, the program will close too