fix(nvim): Neovim can not update the last color scheme
Bug: if the user removes the current color scheme (uninstall t, by
example). The next Neovim session causes an error, and the user can not
restore the last color scheme anymore

Cause: the 'colorscheme.lua' script tries to load the last selected
color scheme. If the user removes it, the color scheme update will throw
an error and exit the script. So the auto-command to save the selected
color scheme will not be loaded

Solution: load the auto-command that saves the current color scheme
before applying it. Set the current color scheme with `pcall()` and show
a notification if Neovim can not load it