fix(nvim): error message of lazy loaded Telescope picker
Bug: If the user forgot to lazy load a telescope picker with the
`MYPLUGFUNC.load_telescope_extension` function, my configuration throws
an error message. The current bug causes an error with 'noice.nvim' and
this warning feature

Cause: My notification backend is 'nvim-notify'. It accepts multi-line
strings as the message content, but 'noice.nvim' does not. The message
thrown if the user forgot to lazy load a picker uses a multi-line string
that does not work with 'noice.nvim'

Solution: Use a single string in the body of the notification message