refactor(nvim): 'nvim-ufo' key maps
Keymaps for moving through fondings do not apply to insert mode. Only
for normal and visual modes.

Changed key maps:

`<A-H>` remapped to zh: Fold level -1
`<A-L>` remapped to zl: Fold level +1

`<A-j>` remapped to `<A-J>`: Next fold
`<A-k>` remapped to `<A-K>`: Previous fold
`<A-h>` remapped to `<A-H>`: Close previous fold
`<A-l>` remapped to `<A-L>`: Open current fold