fix(Makefile): can not run the 'once_before_disable' scripts
Bug: do not run the 'once_before_disable' script after the user disable
some package

Reasons: the command to execute the script uses the package name to
check if the script is of 'disable' or 'enable' type. The correct
behavior is to use the script folder. This comparison checks by the
'disabled' keyword, but no script folder has this keyword. The correct
keyword is 'disable'

Solution: check the script type by the 'disable' keyword. Check the
script type by the script folder instead of the package name