fix(nvim): can not create 'LspFileTypes' user command
Bug: 'LspFileTypes' command is not loaded when 'vue_ls' is installed
with Mason

Reason: My LSP server configuration ends the server configuration if
there is no lspconfig configuration available. This is done with a
'return' statement, finalizing not only the server configuration but
also the 'config' function of the lspconfig plugin before the
'LspFileTypes' user command is created

Solution: place the configuration code inside a function. The 'return'
statement ends only the LSP server configuration, not the 'config'
function