fix(clipboard): does not clear the clipboard if the process is killed
Bug: if the user kills the process of the script that clears the
clipboard, the clipboard is not cleared

Reason: this is a security issue, but I did not thought of it before

Solution: add the '__clear_now.sh' script to clear the clipboard. Add a
`trap` command in the '__clear.sh' script to automatic clear the
clipboard if the process is killed or an error occurs. Also update the
'copy.sh' script to use this new feature