fix(nvim): `MYFUNC.iter_path()` only checks the first directoy
Bug: `MYFUNC.iter_path()` only checks the first directoy and ignore the
other directories. Returning `nil` if the first directory is does not
matches the filter conditions

Reason: the function exits if the directory is different from '/'.
Because this is the usual case, the function exists after the first
directory

Solution: change the check to exit if the directory is equal to '/'