feat(nvim/my_scripts/indent): highlight incomplete indentation levels
Before: the indentation highlight doesn't highlight incomplete
indentation levels. The following code is a example that would not
highlight in a file with 4 spaces of indentation because has only 2
spaces of indentation:

```cpp
  private:
```

Now: the indentation highlight highlights incomplete indentation levels.
The code example above would highlight even in a file with 4 spaces of
indentation