fix(zsh/prompt): open the prompt with a small width causes CPU stress
Bug: opening a terminal with a small number of columns (width) results
in a heavy CPU load

Cause: some width computations depends on the number of columns. If it
is too small, a negative buffer-overflow occurs resulting in a big
number (order: 2^32). This value is used to print characters. Resume:
the buffer overflow makes the prompt output millions of characters

Solution: Checks by the buffer overflow before it occurs. Also check the
number of characters to print and disable it if it is too big (bigger
than the number of columns)