feat(profile): activate mise in 'shims' mode
In non-shims mode, mise adds the binaries paths to the user PATH and
manages it when something changes (example: the user changes form Go 20
to Go 21). This setup works when in interactive shells, because mise
adds hooks to manage the user PATH

When sourcing the profile file, mise can not add hooks, so the better
solution is to activate it in shims mode. Instead of managing the PATH,
mise adds all the binaries to a separate directory and just adds this
single directory to the PATH. No need to change the PATH in hooks

NOTE: My ZSH configuration activates mise in non-shims mode. Mise will
attempt to remove the shims directory from the PATH and use hooks to
manage the PATH